move HoverHandler to it's own class, have BibleGetQuoteFrame use VersionsSelect
